Renamed packages with bad naming
