+++ connector-api/src/main/java/it/pagopa/selfcare/party/registry_proxy/connector/api/NationalRegistriesConnector.java
@@ -1,8 +1,8 @@
 package it.pagopa.selfcare.party.registry_proxy.connector.api;
 
-import it.pagopa.selfcare.party.registry_proxy.connector.model.nationalregistries.Businesses;
-import it.pagopa.selfcare.party.registry_proxy.connector.model.nationalregistries.LegalAddressResponse;
-import it.pagopa.selfcare.party.registry_proxy.connector.model.nationalregistries.VerifyLegalResponse;
+import it.pagopa.selfcare.party.registry_proxy.connector.model.national_registries.Businesses;
+import it.pagopa.selfcare.party.registry_proxy.connector.model.national_registries.LegalAddressResponse;
+import it.pagopa.selfcare.party.registry_proxy.connector.model.national_registries.VerifyLegalResponse;
 
 public interface NationalRegistriesConnector{
