+++ connector-api/src/main/java/it/pagopa/selfcare/party/registry_proxy/connector/api/PDNDInfoCamereConnector.java
@@ -1,6 +1,6 @@
 package it.pagopa.selfcare.party.registry_proxy.connector.api;
 
-import it.pagopa.selfcare.party.registry_proxy.connector.model.nationalregistriespdnd.PDNDBusiness;
+import it.pagopa.selfcare.party.registry_proxy.connector.model.national_registries_pdnd.PDNDBusiness;
 import java.util.List;
 
 public interface PDNDInfoCamereConnector {
