+++ connector-api/src/main/java/it/pagopa/selfcare/party/registry_proxy/connector/model/national_registries/LegalAddressFilter.java
@@ -1,4 +1,4 @@
-package it.pagopa.selfcare.party.registry_proxy.connector.model.nationalregistries;
+package it.pagopa.selfcare.party.registry_proxy.connector.model.national_registries;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
