+++ connector-api/src/main/java/it/pagopa/selfcare/party/registry_proxy/connector/model/national_registries_pdnd/PDNDBusiness.java
@@ -1,4 +1,4 @@
-package it.pagopa.selfcare.party.registry_proxy.connector.model.nationalregistriespdnd;
+package it.pagopa.selfcare.party.registry_proxy.connector.model.national_registries_pdnd;
 
 import java.util.List;
 import lombok.Data;
