+++ connector/rest/src/main/java/it/pagopa/selfcare/party/registry_proxy/connector/rest/client/NationalRegistriesRestClient.java
@@ -2,8 +2,8 @@ package it.pagopa.selfcare.party.registry_proxy.connector.rest.client;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
-import it.pagopa.selfcare.party.registry_proxy.connector.model.nationalregistries.Businesses;
-import it.pagopa.selfcare.party.registry_proxy.connector.model.nationalregistries.LegalAddressRequest;
+import it.pagopa.selfcare.party.registry_proxy.connector.model.national_registries.Businesses;
+import it.pagopa.selfcare.party.registry_proxy.connector.model.national_registries.LegalAddressRequest;
 import it.pagopa.selfcare.party.registry_proxy.connector.rest.config.NationalRegistriesRestClientConfig;
 import it.pagopa.selfcare.party.registry_proxy.connector.rest.model.AdELegalOKDto;
 import it.pagopa.selfcare.party.registry_proxy.connector.rest.model.AdELegalRequestBodyDto;
