+++ connector/rest/src/test/java/it/pagopa/selfcare/party/registry_proxy/connector/rest/NationalRegistriesConnectorImplTest.java
@@ -1,6 +1,6 @@
 package it.pagopa.selfcare.party.registry_proxy.connector.rest;
 
-import it.pagopa.selfcare.party.registry_proxy.connector.model.nationalregistries.*;
+import it.pagopa.selfcare.party.registry_proxy.connector.model.national_registries.*;
 import it.pagopa.selfcare.party.registry_proxy.connector.rest.client.NationalRegistriesRestClient;
 import it.pagopa.selfcare.party.registry_proxy.connector.rest.model.AdELegalOKDto;
 import it.pagopa.selfcare.party.registry_proxy.connector.constant.AdEResultCodeEnum;
