+++ web/src/main/java/it/pagopa/selfcare/party/registry_proxy/web/model/mapper/BusinessMapper.java
@@ -1,7 +1,7 @@
 package it.pagopa.selfcare.party.registry_proxy.web.model.mapper;
 
 
-import it.pagopa.selfcare.party.registry_proxy.connector.model.nationalregistries.Business;
+import it.pagopa.selfcare.party.registry_proxy.connector.model.national_registries.Business;
 import it.pagopa.selfcare.party.registry_proxy.web.model.BusinessResource;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
