+++ web/src/main/java/it/pagopa/selfcare/party/registry_proxy/web/model/mapper/PDNDInfoCamereBusinessMapper.java
@@ -1,6 +1,6 @@
 package it.pagopa.selfcare.party.registry_proxy.web.model.mapper;
 
-import it.pagopa.selfcare.party.registry_proxy.connector.model.nationalregistriespdnd.PDNDBusiness;
+import it.pagopa.selfcare.party.registry_proxy.connector.model.national_registries_pdnd.PDNDBusiness;
 import it.pagopa.selfcare.party.registry_proxy.web.model.PDNDBusinessResource;
 import org.mapstruct.Mapper;
