+++ web/src/test/java/it/pagopa/selfcare/party/registry_proxy/web/model/mapper/BusinessesMapperTest.java
@@ -3,8 +3,8 @@ package it.pagopa.selfcare.party.registry_proxy.web.model.mapper;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
-import it.pagopa.selfcare.party.registry_proxy.connector.model.nationalregistries.Business;
-import it.pagopa.selfcare.party.registry_proxy.connector.model.nationalregistries.Businesses;
+import it.pagopa.selfcare.party.registry_proxy.connector.model.national_registries.Business;
+import it.pagopa.selfcare.party.registry_proxy.connector.model.national_registries.Businesses;
 import it.pagopa.selfcare.party.registry_proxy.web.model.BusinessesResource;
 
 import java.util.ArrayList;
